# sourceCode/src/main/java/com/company/payment/payment_backend/model/dto/CaptureResponse.java
package com.company.payment.payment_backend.model.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

public class CaptureResponse {
    public CaptureResponse() {
    }

    public String getOrderId() {
        return orderId;
    }

    public CaptureResponse(String orderId, String status, String message) {
        this.orderId = orderId;
        this.status = status;
        this.message = message;
    }

    public void setOrderId(String orderId) {
        this.orderId = orderId;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    private String orderId;
    private String status;
    private String message;
}